Adding serialBone()
The serialBone is a bone that can be used for serial numbers, codes, or other, related content having an optional minimum or maximum length. Minimum length can be filled up with a fill-char and aligned left or right.

There is a related branch in ViUR vi to handle this kind of bone.