Adding filterRelationalBone()
This kind of bone is working like an usual relationalBone, except that it has the capability to use a list or run a callback-function that returns a list to retrieve reference keys whose entries are permitted. Entries that do not fit into this set are filtered out when read, and inserted in when written, so that data remains consistent.