New viur.bone.relational.notFoundStrategy config
This config switch allow to run different strategies on how valid, but non-existing keys to references should be handled by the ViUR system. Default is "raise", but "report" or "ignore" can also be used.

Currently this switch only affects when a server module is manually calling skel.setBoneValue(), so fromClient() still fails with invalid data.

Can it be useful to enhance fromClient also to the notFoundStrategy-switch?